BLS12-381 and BLS12-377 pairing works using common code
