+++ libff/CMakeLists.txt
@@ -41,8 +41,6 @@ add_library(
   algebra/curves/mnt753/mnt4753/mnt4753_pp.cpp
   algebra/curves/bls12_381/bls12_381_g1.cpp
   algebra/curves/bls12_381/bls12_381_g2.cpp
-  algebra/curves/bls12_381/bls12_381_init.cpp
-  algebra/curves/bls12_381/bls12_381_pairing.cpp
   algebra/curves/bls12_381/bls12_381_pp.cpp
   algebra/curves/sw6_bis/sw6_bis_g1.cpp
   algebra/curves/sw6_bis/sw6_bis_g2.cpp
@@ -57,7 +55,6 @@ add_library(
   algebra/curves/bls12_377/bls12_377_g1.cpp
   algebra/curves/bls12_377/bls12_377_g2.cpp
   algebra/curves/bls12_377/bls12_377_init.cpp
-  algebra/curves/bls12_377/bls12_377_pairing.cpp
   algebra/curves/bls12_377/bls12_377_pp.cpp
   algebra/curves/alt_bn128/alt_bn128_g1.cpp
   algebra/curves/alt_bn128/alt_bn128_g2.cpp
