+++ libff/algebra/curves/bls12_377/bls12_377_init.cpp
@@ -1,6 +1,7 @@
 #include <libff/algebra/curves/bls12_377/bls12_377_g1.hpp>
 #include <libff/algebra/curves/bls12_377/bls12_377_g2.hpp>
 #include <libff/algebra/curves/bls12_377/bls12_377_init.hpp>
+#include <libff/algebra/curves/bls12_377/bls12_377_pp.hpp>
 
 namespace libff {
 
@@ -15,11 +16,7 @@ bls12_377_Fq bls12_377_twist_mul_by_b_c1;
 bls12_377_Fq2 bls12_377_twist_mul_by_q_X;
 bls12_377_Fq2 bls12_377_twist_mul_by_q_Y;
 
-bigint<bls12_377_q_limbs> bls12_377_ate_loop_count;
-bool bls12_377_ate_is_loop_count_neg;
-bigint<12*bls12_377_q_limbs> bls12_377_final_exponent;
-bigint<bls12_377_q_limbs> bls12_377_final_exponent_z;
-bool bls12_377_final_exponent_is_z_neg;
+bls12_377_pp::Fqe_type bls12_377_pp::TWIST_COEFF_B;
 
 void init_bls12_377_params()
 {
@@ -78,7 +75,7 @@ void init_bls12_377_params()
     bls12_377_Fq::t_minus_1_over_2 = bigint_q("1837921289030710838195067919506396475074392872918698035817074744121558668640693829665401097909504529");
     bls12_377_Fq::multiplicative_generator = bls12_377_Fq("15");
     bls12_377_Fq::root_of_unity = bls12_377_Fq("146552004846884389553264564610149105174701957497228680529098805315416492923550540437026734404078567406251254115855");
-    bls12_377_Fq::nqr = bls12_377_Fq("5");
+    bls12_377_Fq::nqr = bls12_377_Fq("258664426012969094010652733694893533536393512754914660539884262666720468348340822774968888139573360124440321458172"); // -5
     bls12_377_Fq::nqr_to_t = bls12_377_Fq("32863578547254505029601261939868325669770508939375122462904745766352256812585773382134936404344547323199885654433");
    // Done with Fq
 
@@ -95,7 +92,7 @@ void init_bls12_377_params()
     // Done with Fq2
 
     /* parameters for Fq6 */
-    bls12_377_Fq6::non_residue = bls12_377_Fq2(bls12_377_Fq("0"),bls12_377_Fq("1"));
+    bls12_377_Fq6::non_residue = bls12_377_Fq2::nqr; //bls12_377_Fq2(bls12_377_Fq("0"),bls12_377_Fq("1"));
     bls12_377_Fq6::Frobenius_coeffs_c1[0] = bls12_377_Fq2(bls12_377_Fq("1"),bls12_377_Fq("0"));
     bls12_377_Fq6::Frobenius_coeffs_c1[1] = bls12_377_Fq2(bls12_377_Fq("80949648264912719408558363140637477264845294720710499478137287262712535938301461879813459410946"),bls12_377_Fq("0"));
     bls12_377_Fq6::Frobenius_coeffs_c1[2] = bls12_377_Fq2(bls12_377_Fq("80949648264912719408558363140637477264845294720710499478137287262712535938301461879813459410945"),bls12_377_Fq("0"));
@@ -112,7 +109,7 @@ void init_bls12_377_params()
 
 
     /* parameters for Fq12 */
-    bls12_377_Fq12::non_residue = bls12_377_Fq2(bls12_377_Fq("0"),bls12_377_Fq("1"));
+    bls12_377_Fq12::non_residue = bls12_377_Fq2::nqr; //bls12_377_Fq2(bls12_377_Fq("0"),bls12_377_Fq("1"));
     bls12_377_Fq12::Frobenius_coeffs_c1[0] = bls12_377_Fq2(bls12_377_Fq("1"),bls12_377_Fq("0"));
     bls12_377_Fq12::Frobenius_coeffs_c1[1] = bls12_377_Fq2(bls12_377_Fq("92949345220277864758624960506473182677953048909283248980960104381795901929519566951595905490535835115111760994353"),bls12_377_Fq("0"));
     bls12_377_Fq12::Frobenius_coeffs_c1[2] = bls12_377_Fq2(bls12_377_Fq("80949648264912719408558363140637477264845294720710499478137287262712535938301461879813459410946"),bls12_377_Fq("0"));
@@ -131,7 +128,7 @@ void init_bls12_377_params()
     /* choice of short Weierstrass curve and its twist */
     bls12_377_coeff_b = bls12_377_Fq("1");
     bls12_377_twist = bls12_377_Fq2(bls12_377_Fq("0"), bls12_377_Fq("1"));
-    bls12_377_twist_coeff_b = bls12_377_Fq2(bls12_377_Fq("0"), bls12_377_Fq("155198655607781456406391640216936120121836107652948796323930557600032281009004493664981332883744016074664192874906"));
+    bls12_377_pp::TWIST_COEFF_B = bls12_377_twist_coeff_b = bls12_377_Fq2(bls12_377_Fq("0"), bls12_377_Fq("155198655607781456406391640216936120121836107652948796323930557600032281009004493664981332883744016074664192874906"));
     bls12_377_twist_mul_by_b_c0 = bls12_377_coeff_b * bls12_377_Fq2::non_residue;
     bls12_377_twist_mul_by_b_c1 = bls12_377_coeff_b * bls12_377_Fq2::non_residue;
     bls12_377_twist_mul_by_q_X = bls12_377_Fq2(bls12_377_Fq("80949648264912719408558363140637477264845294720710499478137287262712535938301461879813459410946"),
@@ -274,20 +271,5 @@ void init_bls12_377_params()
     // bls12_377_G2::fixed_base_exp_window_table.push_back(0);
     // // window 22 is unbeaten in [31673814.95, inf]
     // bls12_377_G2::fixed_base_exp_window_table.push_back(31673815);
-
-
-
-    /* pairing parameters */
-    // TODO: final_exp_z=last_chunk? 
-    /* 
-     * https://eprint.iacr.org/2017/1174.pdf
-     * ate_loop_count=t where t in the value chosen in q(t) and r(t) parameterization of BLS curve.
-     * for BLS12_377 t=3·2^46·(7·13·499)+1 s.t. t=1 (mod 3·2^46) to have a high 2-adicity for q and r. 
-    */
-    bls12_377_ate_loop_count = bigint_q("9586122913090633729");
-    bls12_377_ate_is_loop_count_neg = false; // true in BLS12_381 for instance
-    bls12_377_final_exponent = bigint<12*bls12_377_q_limbs>("10623521018019860488254031663707568428798032905123811199571213965079129114663661236359849629341526275899063345613340067081670062620727617884137487754739150147491204559514205186492385590272208934467461444944652711005169371168250068790820776124772095630237102189827733019989835063334551453893534663070786533932633573962932272563471643288531959637300817070265537429506484880990981069041269405383502889677357082012807298529931118124428569059822346289745077401570134157444973271520981774047146918354408632568723153146248333028827919406785654402107153546667815607201488590832478225403444136409349877481268154817904541340614173261949772403060924324366861723245182619859389254985008236007465814273361497134138868945580557938161335670207544906643574043606819537336472235809927599628123275314288006170804044560238676463931639339711913111080974582593228138704154320599775683095604041309000197025419968125718018311805959315220036948621879242495199408833915486421612374480018459896018440926235261824654956932384859260479372776022979736734221629097297890154692194441528462770218811795624471108972377573690833913231260547835550851256817740247389770320334698430697237343583761719223414894063451411431859122738488311580005412765070251810159991897110936324943232526870280724876946523218213525646968094720");
-    bls12_377_final_exponent_z = bigint_q("9586122913090633729");
-    bls12_377_final_exponent_is_z_neg = false;
 }
 } // libff
