+++ libff/algebra/curves/bls12_377/bls12_377_init.hpp
@@ -42,11 +42,13 @@ extern bls12_377_Fq2 bls12_377_twist_mul_by_q_X;
 extern bls12_377_Fq2 bls12_377_twist_mul_by_q_Y;
 
 // parameters for pairing
+/*
 extern bigint<bls12_377_q_limbs> bls12_377_ate_loop_count;
 extern bool bls12_377_ate_is_loop_count_neg;
 extern bigint<12*bls12_377_q_limbs> bls12_377_final_exponent;
 extern bigint<bls12_377_q_limbs> bls12_377_final_exponent_z;
 extern bool bls12_377_final_exponent_is_z_neg;
+*/
 
 void init_bls12_377_params();
