+++ libff/algebra/fields/fp6_3over2.tcc
@@ -171,31 +171,15 @@ void Fp6_3over2_model<n,modulus>::multiply_by_c1(const Fp6_3over2_model<n,modulu
 template<mp_size_t n, const bigint<n>& modulus>
 void Fp6_3over2_model<n,modulus>::multiply_by_c01(const my_Fp6 &a, const my_Fp2 &c0, const my_Fp2 & c1)
 {
-    my_Fp2 tmp1;
-    my_Fp2 tmp2;
-    my_Fp2 tmp3;
-    tmp1.add(a.c1, a.c2);
-    tmp2.add(a.c0, a.c1);
-    tmp3.add(a.c0, a.c2);
+    const my_Fp2 tmp2 = a.c0 + a.c1;
+    const my_Fp2 tmp3 = a.c0 + a.c2;
 
-    my_Fp2 a_a;
-    my_Fp2 b_b;
-    a_a.multiply(a.c0, c0);
-    b_b.multiply(a.c1, c1);
+    const my_Fp2 a_a = c0 * a.c0;
+    const my_Fp2 b_b = c1 * a.c1;
 
-    this->c0.multiply(c1, tmp1);
-    this->c0.subtract(this->c0, b_b);
-    this->c0.multiply_by_nonresidue(this->c0);
-    this->c0.add(this->c0, a_a);
-
-    this->c1.add(c0, c1);
-    this->c1.multiply(this->c1, tmp2);
-    this->c1.subtract(this->c1, a_a);
-    this->c1.subtract(this->c1, b_b);
-
-    this->c2.multiply(c0, tmp3);
-    this->c2.subtract(this->c2, a_a);
-    this->c2.add(this->c2, b_b);
+    this->c0 = ((c1 * (a.c1 + a.c2)) - b_b).multiply_by_nonresidue() + a_a;
+    this->c1 = ((c0 + c1) * tmp2) - a_a - b_b;
+    this->c2 = (c0 * tmp3) - a_a + b_b;
 }
 
 template<mp_size_t n, const bigint<n>& modulus>
